Rename TestComponenet to MockComponent and add test package for unit tests
